pyproject.toml
[tool.poetry]
name = "bookstore_api"
version = "0.1.0"
description = ""
authors = ["Oluwaseun Ilesanmi <seunowonikoko@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.2"
python-dotenv = "^1.2.1"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.1.0"
psycopg2-binary = "^2.9.11"
sqlalchemy = "^2.0.44"
alembic = "^1.17.1"
flask-marshmallow = "^1.3.0"
flask-restful = "^0.3.10"
flask-jwt-extended = "^4.7.1"
marshmallow-sqlalchemy = "^1.4.2"
redis = "^7.1.0"
boto3 = "^1.41.5"

[tool.poetry.scripts]
start = "bookstore_api:app.run"
# This requires FLASK_APP environment variable to be set.
my-cli = "flask.cli:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
